src/frontend/requirements.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
dotenv==0.9.9
exceptiongroup==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
nest-asyncio==1.6.0
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
